Move query parsing function from compare component to LocationComparisonComponent
